Move SuperTriggerCell Impl to new class
