+++ L1Trigger/L1THGCal/interface/concentrator/HGCalConcentratorSelectionImpl.h
@@ -19,8 +19,7 @@ class HGCalConcentratorSelectionImpl
     HGCalConcentratorSelectionImpl(const edm::ParameterSet& conf);
 
     void bestChoiceSelectImpl(const std::vector<l1t::HGCalTriggerCell>& trigCellVecInput, std::vector<l1t::HGCalTriggerCell>& trigCellVecOutput);    
-    void thresholdSelectImpl(const std::vector<l1t::HGCalTriggerCell>& trigCellVecInput, std::vector<l1t::HGCalTriggerCell>& trigCellVecOutput);  
-    void superTriggerCellSelectImpl(uint32_t module_id, const std::vector<l1t::HGCalTriggerCell>& trigCellVecInput, std::vector<l1t::HGCalTriggerCell>& trigCellVecOutput);
+    void thresholdSelectImpl(const std::vector<l1t::HGCalTriggerCell>& trigCellVecInput, std::vector<l1t::HGCalTriggerCell>& trigCellVecOutput);      
   
     // Retrieve parameters
     size_t   nCellsInModule() const {return nCellsInModule_;}
@@ -34,29 +33,6 @@ class HGCalConcentratorSelectionImpl
     double   TCThresholdBH_MIP() const {return TCThresholdBH_MIP_;} 
 
   private:
-
-    int getSuperTriggerCellId_(int detid) const ;
-
-    struct SuperTriggerCell {
-        float sumPt, sumMipPt;
-        int sumHwPt, maxHwPt; 
-        unsigned maxId;
-        SuperTriggerCell() : sumPt(0), sumMipPt(0), sumHwPt(0), maxHwPt(0), maxId(0) {}
-        void add(const l1t::HGCalTriggerCell &c) {
-            sumPt += c.pt();
-            sumMipPt += c.mipPt();
-            sumHwPt += c.hwPt();
-            if (maxId == 0 || c.hwPt() > maxHwPt) {
-                maxHwPt = c.hwPt();
-                maxId = c.detId();
-            }
-        }
-        void assignEnergy(l1t::HGCalTriggerCell &c) const {
-            c.setHwPt(sumHwPt);
-            c.setMipPt(sumMipPt);
-            c.setP4(math::PtEtaPhiMLorentzVector(sumPt, c.eta(), c.phi(), c.mass())); // there's no setPt
-        }
-    };
 
     size_t   nData_;
     size_t   nCellsInModule_;
